refactor: decoupled demos from hashing
