+++ demos/hashing/hashing_algorithms_demo.c
